requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
dill==0.3.6
Django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-froala-editor==4.0.10
django-modeltranslation==0.18.5
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
lxml==4.9.1
Markdown==3.4.1
multiprocess==0.70.14
pathos==0.3.0
Pillow==9.2.0
pox==0.3.2
ppft==1.7.6.6
pyaes==1.6.1
pycparser==2.21
PyExecJS==1.5.1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
translators==5.5.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
whitenoise==6.2.0
